feat: Unified event system, capability security, fault recovery
Major architectural improvements across kernel and userspace:

## Unified Event System (BSD kqueue-inspired)
- Multiple timers per task (8 slots) with recurring support
- EventFilter enum for type-safe subscriptions
- Batch kevent_wait() for efficiency
- New syscalls: kevent_subscribe, kevent_unsubscribe, kevent_timer, kevent_wait
- Migrated consoled, devd, gpio, logd to kevent API

## Capability-based Security
- exec_with_caps syscall for spawning with explicit capabilities
- Capability presets: BUS_DRIVER, DEVICE_DRIVER, FS_DRIVER, SERVICE_DRIVER
- devd grants appropriate capabilities to each driver type
- Child capabilities limited to parent's (delegation model)
- verify_binary_signature() stub for future crypto verification

## Bus Capabilities Reporting
- Hardware capability flags (PCIE_BUS_MASTER, MSI, LINK_UP, USB_2_0/3_0)
- StateSnapshot.capabilities now reports actual hardware state

## Fault Recovery
- Userspace faults terminate task gracefully (not system hang)
- Children killed, parent notified via ChildExit event
- Exit codes reflect signal (-11=SIGSEGV, -7=SIGBUS, -6=SIGABRT)
- devd crash recovery: kill all tasks, reset buses, respawn devd

## Driver Restart Improvements
- Restart waits for bus Safe signal, not just timer
- auto_started flag: only auto-restart drivers spawned by devd
- Manually started drivers (shell) not auto-restarted

## Shell Improvements
- Terminal size detection via cursor position report
- Periodic re-check every 5 seconds, manual resize command
- Fixed echo not working (insert_char bug in readline.rs)
- Object-oriented output for ls/ps (table-like display)

## Logging & Tracing
- Structured logging: <ms> <LVL> <subsys> [ctx] <event> key=val
- Migrated all drivers to ulog (except wifid partial)
- Added logd daemon for log distribution
- Tracing framework with span support

## Code Cleanup
- Removed dead user/driver/msc/ (usbd handles MSC)
- Removed deprecated syscalls: pci_enumerate, pci_bar_map, pci_claim
- Consolidated wifid drivers (wifid2, wifid3 → wifid)
- Documentation restructured into docs/ hierarchy

+++ src/kernel/uaccess.rs
@@ -65,7 +65,7 @@ pub enum UAccessError {
 
 impl UAccessError {
     /// Convert to errno-style error code
-    pub fn to_errno(self) -> i64 {
+    pub fn to_errno(self) -> i32 {
         match self {
             UAccessError::NullPointer => -14,      // EFAULT
             UAccessError::KernelAddress => -14,    // EFAULT
